+++ test/unit/json_validation_test.rb
@@ -1,22 +1,23 @@
 require_relative '../test_helper'
 
 class JsonValidationTest < Test::Unit::TestCase
- 
+
   # turn off the ridiculous warnings
   $VERBOSE=nil
 
-  ERROR_DIR = File.join('errors', 'JsonValidationTest')
-  EXAMPLE_ROOT = File.join('examples','json')
+  ERROR_DIR = File.join('tmp', 'errors', 'JsonValidationTest')
+  EXAMPLE_ROOT = File.join('lib', 'fhir_models', 'examples', 'json')
 
   # Automatically generate one test method per example file
   example_files = File.join(EXAMPLE_ROOT, '**', '*.json')
+  raise 'No Examples Found' if Dir[example_files].empty?
 
   # Create a blank folder for the errors
   FileUtils.rm_rf(ERROR_DIR) if File.directory?(ERROR_DIR)
   FileUtils.mkdir_p ERROR_DIR
 
   Dir.glob(example_files).each do | example_file |
-    example_name = File.basename(example_file, ".json")
+    example_name = File.basename(example_file, '.json')
     define_method("test_json_validation_#{example_name}") do
       run_json_validation_test(example_file, example_name)
     end
@@ -28,9 +29,9 @@ class JsonValidationTest < Test::Unit::TestCase
     errors = resource.validate
     if !errors.empty?
       File.open("#{ERROR_DIR}/#{example_name}.err", 'w:UTF-8') {|file| file.write(JSON.pretty_unparse(errors))}
-      File.open("#{ERROR_DIR}/#{example_name}.json", 'w:UTF-8') {|file| file.write(input_json)}      
+      File.open("#{ERROR_DIR}/#{example_name}.json", 'w:UTF-8') {|file| file.write(input_json)}
     end
-    assert errors.empty?, "Resource failed to validate."
+    assert errors.empty?, 'Resource failed to validate.'
   end
 
 end
